perf(plugins): keep provider discovery metadata lightweight
