+++ config/assertion-safety-baseline.txt
@@ -520,7 +520,7 @@ extensions/google/onboard.ts	3
 extensions/google/provider-registration.ts	2
 extensions/google/realtime-voice-provider.ts	6
 extensions/google/transport-stream.ts	13
-extensions/google/vertex-adc.ts	6
+extensions/google/vertex-adc.ts	4
 extensions/google/video-generation-provider.ts	13
 extensions/googlechat/src/accounts.ts	1
 extensions/googlechat/src/actions.ts	1
