+++ extensions/anthropic-vertex/region.adc.test.ts
@@ -1,7 +1,7 @@
 // Anthropic Vertex tests cover region.adc plugin behavior.
 import { platform } from "node:os";
 import path from "node:path";
-import { afterAll, afterEach, describe, expect, it, vi } from "vitest";
+import { afterAll, beforeEach, describe, expect, it, vi } from "vitest";
 
 const { existsSyncMock, readFileSyncMock } = vi.hoisted(() => ({
   existsSyncMock: vi.fn(),
@@ -10,20 +10,12 @@ const { existsSyncMock, readFileSyncMock } = vi.hoisted(() => ({
 
 vi.mock("node:fs", async () => {
   const actual = await vi.importActual<typeof import("node:fs")>("node:fs");
-  existsSyncMock.mockImplementation((pathname) => actual.existsSync(pathname));
-  readFileSyncMock.mockImplementation((pathname, options) =>
-    String(pathname) === "/tmp/vertex-adc.json"
-      ? '{"project_id":"vertex-project"}'
-      : actual.readFileSync(pathname, options as never),
-  );
   return {
     ...actual,
     existsSync: existsSyncMock,
-    readFileSync: readFileSyncMock,
     default: {
       ...actual,
       existsSync: existsSyncMock,
-      readFileSync: readFileSyncMock,
     },
   };
 });
@@ -35,9 +27,16 @@ vi.mock("openclaw/plugin-sdk/secret-file-runtime", () => ({
 import { hasAnthropicVertexAvailableAuth, resolveAnthropicVertexProjectId } from "./region.js";
 
 describe("anthropic-vertex ADC reads", () => {
-  afterEach(() => {
-    existsSyncMock.mockClear();
-    readFileSyncMock.mockClear();
+  beforeEach(() => {
+    existsSyncMock.mockReset();
+    readFileSyncMock.mockReset();
+    // The secret-file fixture owns its contents, independently of unrelated fs imports.
+    readFileSyncMock.mockImplementation((pathname) => {
+      if (String(pathname) !== "/tmp/vertex-adc.json") {
+        throw new Error(`unexpected ADC fixture path: ${String(pathname)}`);
+      }
+      return '{"project_id":"vertex-project"}';
+    });
   });
 
   afterAll(() => {
@@ -72,11 +71,9 @@ describe("anthropic-vertex ADC reads", () => {
     readFileSyncMock.mockImplementation((pathname, options) =>
       String(pathname) === defaultAdcPath
         ? '{"project_id":"vertex-project"}'
-        : String(pathname) === "/tmp/vertex-adc.json"
-          ? '{"project_id":"vertex-project"}'
-          : (() => {
-              throw new Error(`unexpected readFileSync(${String(pathname)}, ${String(options)})`);
-            })(),
+        : (() => {
+            throw new Error(`unexpected readFileSync(${String(pathname)}, ${String(options)})`);
+          })(),
     );
 
     expect(resolveAnthropicVertexProjectId(env)).toBe("vertex-project");
