+++ extensions/anthropic-vertex/region.ts
@@ -5,7 +5,6 @@
 import { homedir, platform } from "node:os";
 import { join } from "node:path";
 import type { GoogleAuthOptions } from "google-auth-library";
-import { resolveProviderEndpoint } from "openclaw/plugin-sdk/provider-http";
 import { tryReadSecretFileSync } from "openclaw/plugin-sdk/secret-file-runtime";
 import {
   normalizeLowercaseStringOrEmpty,
@@ -42,23 +41,6 @@ export function resolveAnthropicVertexProjectId(
     normalizeOptionalSecretInput(env.GOOGLE_CLOUD_PROJECT) ||
     normalizeOptionalSecretInput(env.GOOGLE_CLOUD_PROJECT_ID) ||
     resolveAnthropicVertexProjectIdFromAdc(env)
-  );
-}
-
-/** Extract a Vertex region from a provider base URL when possible. */
-export function resolveAnthropicVertexRegionFromBaseUrl(baseUrl?: string): string | undefined {
-  const endpoint = resolveProviderEndpoint(baseUrl);
-  return endpoint.endpointClass === "google-vertex" ? endpoint.googleVertexRegion : undefined;
-}
-
-/** Resolve the client region from model base URL first, then env fallback. */
-export function resolveAnthropicVertexClientRegion(params?: {
-  baseUrl?: string;
-  env?: NodeJS.ProcessEnv;
-}): string {
-  return (
-    resolveAnthropicVertexRegionFromBaseUrl(params?.baseUrl) ||
-    resolveAnthropicVertexRegion(params?.env)
   );
 }
