+++ extensions/anthropic-vertex/stream-runtime.ts
@@ -25,11 +25,8 @@ import {
 } from "openclaw/plugin-sdk/provider-model-shared";
 import { copyProviderAcceptanceObserver } from "openclaw/plugin-sdk/provider-transport-runtime";
 import { EnvHttpProxyAgent, fetch as undiciFetch } from "undici";
-import {
-  resolveAnthropicVertexAdcCredentials,
-  resolveAnthropicVertexClientRegion,
-  resolveAnthropicVertexProjectId,
-} from "./region.js";
+import { resolveAnthropicVertexClientRegion } from "./region-endpoint.js";
+import { resolveAnthropicVertexAdcCredentials, resolveAnthropicVertexProjectId } from "./region.js";
 
 const GOOGLE_CLOUD_PLATFORM_SCOPE = "https://www.googleapis.com/auth/cloud-platform";
