+++ extensions/deepinfra/onboard.ts
@@ -3,7 +3,7 @@ import {
   createAliasOnlyPresetAppliers,
   type OpenClawConfig,
 } from "openclaw/plugin-sdk/provider-onboard";
-import { DEEPINFRA_DEFAULT_MODEL_REF } from "./provider-models.js";
+import { DEEPINFRA_DEFAULT_MODEL_REF } from "./provider-static-catalog.js";
 
 export function applyDeepInfraConfig(
   cfg: OpenClawConfig,
