+++ extensions/deepinfra/provider-models.pricing.test.ts
@@ -1,7 +1,8 @@
 import { clearLiveCatalogCacheForTests } from "openclaw/plugin-sdk/provider-catalog-live-runtime";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { buildDeepInfraProvider } from "./api.js";
-import { DEEPINFRA_MODEL_CATALOG, discoverDeepInfraModels } from "./provider-models.js";
+import { discoverDeepInfraModels } from "./provider-models.js";
+import { DEEPINFRA_MODEL_CATALOG } from "./provider-static-catalog.js";
 
 const DEEPINFRA_MODELS_URL =
   "https://api.deepinfra.com/v1/openai/models?sort_by=openclaw&filter=with_meta";
