+++ extensions/google/provider-registration.ts
@@ -7,8 +7,8 @@ import { createProviderApiKeyAuthMethod } from "openclaw/plugin-sdk/provider-ent
 import type { ProviderPlugin } from "openclaw/plugin-sdk/provider-model-shared";
 import { normalizeGoogleModelId } from "./model-id.js";
 import { GOOGLE_GEMINI_DEFAULT_MODEL, applyGoogleGeminiModelDefault } from "./onboard.js";
+import { buildGoogleLiveCatalogProvider } from "./provider-catalog-runtime.js";
 import {
-  buildGoogleLiveCatalogProvider,
   buildGoogleStaticCatalogProvider,
   buildGoogleVertexStaticCatalogProvider,
 } from "./provider-catalog.js";
@@ -28,7 +28,7 @@ import {
   createGoogleGenerativeAiTransportStreamFn,
   createGoogleVertexTransportStreamFn,
 } from "./transport-stream.js";
-import { resolveGoogleVertexConfigApiKey } from "./vertex-adc.js";
+import { resolveGoogleVertexConfigApiKey } from "./vertex-adc-config.js";
 
 function normalizeGoogleVideoInput(
   ctx: Parameters<NonNullable<ProviderPlugin["normalizeResolvedModel"]>>[0],
